requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python-headless
numpy
pillow
